requirements.txt
glfw==2.5.6
numpy==1.24.1
Pillow==9.4.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
